\name{adv}

\docType{data}

\alias{adv}

\title{ADV (acoustic-doppler velocimeter) dataset}

\description{This is a sampling of measurements made with a Nortek Vector
    acoustic Doppler velocimeter deployed as part of the St Lawrence Internal
    Wave Experiment (SLEIWEX).  Various identifying features have been
    redacted, and velocities have modified by adding random numbers chosen from
    a normal distribution with standard deviation of 2 percent of signal.}

\usage{data(adv)}

\source{This file came from the SLEIWEX-2008 experiment.}

\seealso{The documentation for \code{adv-class} in the Oce package explains the
    structure of ADV objects, and also outlines the other functions dealing
    with them.}

\author{Dan Kelley}

\keyword{datasets}

