% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaufort.R
\docType{data}
\name{beaufort}
\alias{beaufort}
\title{Beaufort Scale relationship wave height to wind speed}
\usage{
data(beaufort, package="ocedata")
}
\description{
\code{beaufort} is a data frame describing the Beaufort Scale for sea state.  The
contents are \code{wind}, the wind speed in m/s, and \code{height}, wave height in m.
The data were downloaded from Reference 1 on April 20, 2014.
}
\examples{
data(beaufort, package="ocedata")
plot(beaufort$wind, beaufort$height)
}
\references{
\url{https://www.metoffice.gov.uk/weather/marine/guide/beaufortscale.html}
}
