\name{endeavour}

\docType{data}

\alias{endeavour}

\title{Cook's Endeavour cruise track}

\description{Time, longitude, and latitude for the H.M.S. Endeavour cruise,
    1768-1771}

\usage{data(endeavour)}

\source{The data were extracted from an XML file contained within a zipfile
    provided at the Australian government website
    \url{http://data.gov.au/dataset/journal-of-the-h-m-s-endeavour-1768-1771/}.}

\keyword{datasets}
