% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munk.R
\docType{data}
\name{munk}
\alias{munk}
\title{Munk's (1966) temperature profile}
\usage{
data(munk, package="ocedata")
}
\description{
Temperature profile digitized from the left-hand panel of Munk's
Figure 1.
}
\examples{
data(munk, package="ocedata")
plot(munk$temperature, munk$depth,
    ylim=rev(range(munk$depth)),
    xlab="Temperature [degC]", ylab="Depth [km]")

}
\references{
Walter H. Munk, 1966. Abyssal recipes. \emph{Deep-Sea Research}, 13, 707-730.
}
