\docType{class}
\name{oce-class}
\alias{oce-class}
\title{Class to Store oce Data}
\description{This class stores data in the S4 form used by the oce package.}
\section{Slots}{
    \describe{
        \item{\code{data}}{A list, as for \code{oce}.}
        \item{\code{metadata}}{A list, as for \code{oce}.}
        \item{\code{processingLog}}{A list, as for \code{oce}.}
    }
}
\author{Dan Kelley}

