\name{tdr}
\alias{tdr}

\docType{data}

\title{tdr (temperature-data recorder) dataset}

\description{A sample object of class \code{\link[oce]{tdr-class}}.}

\usage{data(tdr)}

\source{The \acronym{SLEIWEX} experiment.}

\references{\url{http://myweb.dal.ca/kelley/SLEIWEX/index.php}}

\seealso{The documentation for \code{\link[oce]{tdr-class}} explains the structure
    of \code{tdr} objects, and also outlines the other functions dealing with
    them.}

\examples{
\dontrun{
library(oce)
data(tdr)
plot(tdr)
}
}

\author{Dan Kelley}

\keyword{datasets}
