\name{wilson}

\docType{data}

\alias{wilson}

\title{Wilson's (1963) table of seafloor spreading}

\description{Wilson's (1963) table of seafloor spreading.}


\usage{data(wilson)}

\source{Data transcribed from Wilson's (1963) Table 1, taking only the
    first 3 columns, and only those rows for which Wilson reported an estimate
    of spreading rate.}

\references{
    J. Tuzo Wilson, 1963. Evidence from islands on the spreading of ocean floors.  \emph{Nature}, 197(4867):536-538.  \url{//dx.doi.org/10.1038/197536a0.}
}

\keyword{datasets}
