% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endeavour.R
\docType{data}
\name{endeavour}
\alias{endeavour}
\title{Cook's Endeavour cruise track}
\usage{
data(endeavour, package="ocedata")
}
\description{
Time, longitude, and latitude for the H.M.S. Endeavour cruise, 1768-1771.
The data were extracted from an XML file contained within a zipfile provided
at Reference 1, last checked March 13, 2022.
}
\examples{
data(endeavour, package="ocedata")
\dontrun{
library(oce)
data(coastlineWorld, package="oce")
mapPlot(coastlineWorld, col="tan")
mapPoints(endeavour$longitude, endeavour$latitude, pch=20, cex=2/3, col=2)
}
}
\references{
\enumerate{
\item \url{https://www.data.gov.au/dataset/7f03ac78-7210-4702-8cd5-f3e373f4097b}
}
}
