\name{munk}

\docType{data}

\alias{munk}

\title{Munk's (1966) temperature profile}

\description{Temperature profile digitized from the left-hand panel of Munk's
    Figure 1.}

\usage{data(munk)}

\examples{
\dontrun{
library(ocedata)
data(munk)
plot(munk$temperature, munk$depth,
     ylim=rev(range(munk$depth)), xlab="Temperature [degC]", ylab="Depth [km]")
}
}

\references{
Walter H. Munk, 1966.  Abyssal recipes.  Deep-Sea Research, 13, 707-730.
}

\keyword{datasets}
