% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{oce-class}
\alias{oce-class}
\title{Base Class for oce Objects}
\description{
This is used to create objects to be used by oce, created here by ocedata.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{A list containing information about the data.}

\item{\code{data}}{A list containing the data.}

\item{\code{processingLog}}{A list containing time-stamped processing steps.}
}}

