% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocedata.R
\docType{package}
\name{ocedata}
\alias{ocedata}
\title{ocedata: Datasets for the oce package}
\description{
The ocedata package provides datasets for use in oceanographic analysis.
}
\details{
Some commonly-used oceanographic datasets are too large to be stored within
an analysis package such as \CRANpkg{oce}, or are of sufficient general
interest to merit inclusion in a separate package.  Such datasets are
provided here, for use by Oceanographers using the \CRANpkg{oce} package or
other more general packages.
}
