\docType{package}

\name{ocedata}

\alias{ocedata}

\alias{ocedata-package}

\title{Datasets for the oce package}

\description{Provides datasets for use in oceanographic analysis.}

\details{Some commonly-used oceanographic datasets are too large to be stored
    within an analysis package such as \code{oce}, or are of sufficient general
    interest to merit inclusion in a separate package.  Such datasets are
    provided here, for use by Oceanographers using the \code{oce} package or
    other more general packages.}

